+++ UnitySDK/Assets/ML-Agents/Scripts/Sensor/WriteAdapter.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 86bad2e6dded4a62853752a1713981f2
+timeCreated: 1572540197
